+++ graph/api/api.go
@@ -1,0 +1,79 @@
+package api
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/kiali/kiali/business"
+	"github.com/kiali/kiali/graph"
+	"github.com/kiali/kiali/graph/config/cytoscape"
+	"github.com/kiali/kiali/graph/telemetry/istio"
+	"github.com/kiali/kiali/log"
+	"github.com/kiali/kiali/prometheus"
+	"github.com/kiali/kiali/prometheus/internalmetrics"
+)
+
+func GraphNamespaces(business *business.Layer, client *prometheus.Client, o graph.Options) (int, interface{}) {
+	// time how long it takes to generate this graph
+	promtimer := internalmetrics.GetGraphGenerationTimePrometheusTimer(o.GetGraphKind(), o.TelemetryOptions.GraphType, o.InjectServiceNodes)
+	defer promtimer.ObserveDuration()
+
+	// Create a 'global' object to store the business. Global only to the request.
+	globalInfo := graph.NewAppenderGlobalInfo()
+	globalInfo.Business = business
+
+	var trafficMap graph.TrafficMap
+	switch o.TelemetryVendor {
+	case graph.VendorIstio:
+		trafficMap = istio.BuildNamespacesTrafficMap(o.TelemetryOptions, client, globalInfo)
+	default:
+		graph.Error(fmt.Sprintf("TelemetryVendor [%s] not supported", o.TelemetryVendor))
+	}
+
+	code, json := generateGraph(trafficMap, o)
+
+	// update metrics
+	internalmetrics.SetGraphNodes(o.GetGraphKind(), o.TelemetryOptions.GraphType, o.InjectServiceNodes, len(trafficMap))
+
+	return code, json
+}
+
+func GraphNode(business *business.Layer, client *prometheus.Client, o graph.Options) (int, interface{}) {
+	if len(o.Namespaces) != 1 {
+		graph.Error(fmt.Sprintf("Node graph does not support the 'namespaces' query parameter or the 'all' namespace"))
+	}
+
+	// time how long it takes to generate this graph
+	promtimer := internalmetrics.GetGraphGenerationTimePrometheusTimer(o.GetGraphKind(), o.TelemetryOptions.GraphType, o.InjectServiceNodes)
+	defer promtimer.ObserveDuration()
+
+	// Create a 'global' object to store the business. Global only to the request.
+	globalInfo := graph.NewAppenderGlobalInfo()
+	globalInfo.Business = business
+
+	trafficMap := istio.BuildNodeTrafficMap(o.TelemetryOptions, client, globalInfo)
+	code, json := generateGraph(trafficMap, o)
+
+	// update metrics
+	internalmetrics.SetGraphNodes(o.GetGraphKind(), o.TelemetryOptions.GraphType, o.InjectServiceNodes, len(trafficMap))
+
+	return code, json
+}
+
+func generateGraph(trafficMap graph.TrafficMap, o graph.Options) (int, interface{}) {
+	log.Tracef("Generating config for [%s] graph...", o.ConfigVendor)
+
+	promtimer := internalmetrics.GetGraphMarshalTimePrometheusTimer(o.GetGraphKind(), o.TelemetryOptions.GraphType, o.InjectServiceNodes)
+	defer promtimer.ObserveDuration()
+
+	var vendorConfig interface{}
+	switch o.ConfigVendor {
+	case graph.VendorCytoscape:
+		vendorConfig = cytoscape.NewConfig(trafficMap, o.ConfigOptions)
+	default:
+		graph.Error(fmt.Sprintf("ConfigVendor [%s] not supported", o.ConfigVendor))
+	}
+
+	log.Tracef("Done generating config for [%s] graph", o.ConfigVendor)
+	return http.StatusOK, vendorConfig
+}
