+++ graph/telemetry/istio/appender/dead_node_test.go
@@ -121,15 +121,12 @@ func TestDeadNode(t *testing.T) {
 	assert.Equal("istio-ingressgateway", ingressNode.Workload)
 	assert.Equal(10, len(ingressNode.Edges))
 
-	globalInfo := GlobalInfo{
-		Business: businessLayer,
-	}
-	namespaceInfo := NamespaceInfo{
-		Namespace: "testNamespace",
-	}
+	globalInfo := graph.NewAppenderGlobalInfo()
+	globalInfo.Business = businessLayer
+	namespaceInfo := graph.NewAppenderNamespaceInfo("testNamespace")
 
 	a := DeadNodeAppender{}
-	a.AppendGraph(trafficMap, &globalInfo, &namespaceInfo)
+	a.AppendGraph(trafficMap, globalInfo, namespaceInfo)
 
 	assert.Equal(9, len(trafficMap))
 
@@ -151,7 +148,7 @@ func TestDeadNode(t *testing.T) {
 	id, _ := graph.Id("testNamespace", "testNoPodsNoTraffic", "testNamespace", "testNoPodsNoTraffic-v1", "testNoPodsNoTraffic", "v1", graph.GraphTypeVersionedApp)
 	noPodsNoTraffic, ok := trafficMap[id]
 	assert.Equal(true, ok)
-	isDead, ok := noPodsNoTraffic.Metadata["isDead"]
+	isDead, ok := noPodsNoTraffic.Metadata[graph.IsDead]
 	assert.Equal(true, ok)
 	assert.Equal(true, isDead)
 
@@ -192,7 +189,7 @@ func testTrafficMap() map[string]*graph.Node {
 
 	n9 := graph.NewNode("testNamespace", "egress.io", "testNamespace", "", "", "", graph.GraphTypeVersionedApp)
 	n9.Metadata["httpIn"] = 0.8
-	n9.Metadata["isServiceEntry"] = "MESH_EXTERNAL"
+	n9.Metadata[graph.IsServiceEntry] = "MESH_EXTERNAL"
 
 	n10 := graph.NewNode("testNamespace", "egress.not.defined", "testNamespace", "", "", "", graph.GraphTypeVersionedApp)
 	n10.Metadata["httpIn"] = 0.8
