+++ graph/telemetry/istio/appender/unused_node_test.go
@@ -32,7 +32,7 @@ func TestNonTrafficScenario(t *testing.T) {
 	assert.Equal("customer-v1", n.Workload)
 	assert.Equal("customer", n.App)
 	assert.Equal("v1", n.Version)
-	assert.Equal(true, n.Metadata["isUnused"])
+	assert.Equal(true, n.Metadata[graph.IsUnused])
 
 	id, _ = graph.Id("testNamespace", "preference", "testNamespace", "preference-v1", "preference", "v1", a.GraphType)
 	n, ok = trafficMap[id]
@@ -40,7 +40,7 @@ func TestNonTrafficScenario(t *testing.T) {
 	assert.Equal("preference-v1", n.Workload)
 	assert.Equal("preference", n.App)
 	assert.Equal("v1", n.Version)
-	assert.Equal(true, n.Metadata["isUnused"])
+	assert.Equal(true, n.Metadata[graph.IsUnused])
 
 	id, _ = graph.Id("testNamespace", "recommendation", "testNamespace", "recommendation-v1", "recommendation", "v1", a.GraphType)
 	n, ok = trafficMap[id]
@@ -48,7 +48,7 @@ func TestNonTrafficScenario(t *testing.T) {
 	assert.Equal("recommendation-v1", n.Workload)
 	assert.Equal("recommendation", n.App)
 	assert.Equal("v1", n.Version)
-	assert.Equal(true, n.Metadata["isUnused"])
+	assert.Equal(true, n.Metadata[graph.IsUnused])
 
 	id, _ = graph.Id("testNamespace", "recommendation", "testNamespace", "recommendation-v2", "recommendation", "v2", a.GraphType)
 	n, ok = trafficMap[id]
@@ -56,7 +56,7 @@ func TestNonTrafficScenario(t *testing.T) {
 	assert.Equal("recommendation-v2", n.Workload)
 	assert.Equal("recommendation", n.App)
 	assert.Equal("v2", n.Version)
-	assert.Equal(true, n.Metadata["isUnused"])
+	assert.Equal(true, n.Metadata[graph.IsUnused])
 }
 
 func TestOneNodeTrafficScenario(t *testing.T) {
@@ -87,7 +87,7 @@ func TestOneNodeTrafficScenario(t *testing.T) {
 	assert.Equal("customer-v1", n.Workload)
 	assert.Equal("customer", n.App)
 	assert.Equal("v1", n.Version)
-	assert.Equal(nil, n.Metadata["isUnused"])
+	assert.Equal(nil, n.Metadata[graph.IsUnused])
 
 	id, _ = graph.Id("testNamespace", "preference", "testNamespace", "preference-v1", "preference", "v1", a.GraphType)
 	n, ok = trafficMap[id]
@@ -95,7 +95,7 @@ func TestOneNodeTrafficScenario(t *testing.T) {
 	assert.Equal("preference-v1", n.Workload)
 	assert.Equal("preference", n.App)
 	assert.Equal("v1", n.Version)
-	assert.Equal(true, n.Metadata["isUnused"])
+	assert.Equal(true, n.Metadata[graph.IsUnused])
 
 	id, _ = graph.Id("testNamespace", "recommendation", "testNamespace", "recommendation-v1", "recommendation", "v1", a.GraphType)
 	n, ok = trafficMap[id]
@@ -103,7 +103,7 @@ func TestOneNodeTrafficScenario(t *testing.T) {
 	assert.Equal("recommendation-v1", n.Workload)
 	assert.Equal("recommendation", n.App)
 	assert.Equal("v1", n.Version)
-	assert.Equal(true, n.Metadata["isUnused"])
+	assert.Equal(true, n.Metadata[graph.IsUnused])
 
 	id, _ = graph.Id("testNamespace", "recommendation", "testNamespace", "recommendation-v2", "recommendation", "v2", a.GraphType)
 	n, ok = trafficMap[id]
@@ -111,7 +111,7 @@ func TestOneNodeTrafficScenario(t *testing.T) {
 	assert.Equal("recommendation-v2", n.Workload)
 	assert.Equal("recommendation", n.App)
 	assert.Equal("v2", n.Version)
-	assert.Equal(true, n.Metadata["isUnused"])
+	assert.Equal(true, n.Metadata[graph.IsUnused])
 }
 
 func TestVersionWithNoTrafficScenario(t *testing.T) {
@@ -141,7 +141,7 @@ func TestVersionWithNoTrafficScenario(t *testing.T) {
 	assert.Equal("customer", customer.App)
 	assert.Equal("v1", customer.Version)
 	assert.Equal(float64(0.8), unknown.Edges[0].Metadata["http"])
-	assert.Equal(nil, customer.Metadata["isUnused"])
+	assert.Equal(nil, customer.Metadata[graph.IsUnused])
 	assert.Equal(1, len(customer.Edges))
 
 	e := customer.Edges[0]
@@ -150,7 +150,7 @@ func TestVersionWithNoTrafficScenario(t *testing.T) {
 	assert.Equal("preference", preference.App)
 	assert.Equal("v1", preference.Version)
 	assert.Equal(float64(0.8), e.Metadata["http"])
-	assert.Equal(nil, preference.Metadata["isUnused"])
+	assert.Equal(nil, preference.Metadata[graph.IsUnused])
 	assert.Equal(1, len(preference.Edges))
 
 	e = preference.Edges[0]
@@ -159,7 +159,7 @@ func TestVersionWithNoTrafficScenario(t *testing.T) {
 	assert.Equal("recommendation", recommendationV1.App)
 	assert.Equal("v1", recommendationV1.Version)
 	assert.Equal(float64(0.8), e.Metadata["http"])
-	assert.Equal(nil, recommendationV1.Metadata["isUnused"])
+	assert.Equal(nil, recommendationV1.Metadata[graph.IsUnused])
 }
 
 func mockWorkloads() []models.WorkloadListItem {
